Finish cleaning up CA cert saving state machines

+++ src/subproc.c
@@ -161,16 +161,19 @@ void
 cm_subproc_done(struct cm_store_entry *entry, struct cm_subproc_state *state)
 {
 	pid_t pid;
-	if (state->pid != -1) {
-		kill(state->pid, SIGKILL);
-		do {
-			pid = waitpid(state->pid, &state->status, 0);
-		} while ((pid == -1) && (errno == EINTR));
+
+	if (state != NULL) {
+		if (state->pid != -1) {
+			kill(state->pid, SIGKILL);
+			do {
+				pid = waitpid(state->pid, &state->status, 0);
+			} while ((pid == -1) && (errno == EINTR));
+		}
+		if (state->fd != -1) {
+			close(state->fd);
+		}
+		talloc_free(state);
 	}
-	if (state->fd != -1) {
-		close(state->fd);
-	}
-	talloc_free(state);
 }
 
 /* Check if we're done (return 0), or need to be called again (-1). */
